requirements.txt
PySide6
pypdf
PyMuPDF
pdf2docx
pdfplumber
pandas
openpyxl
python-pptx
docx2pdf
Pillow
comtypes
